refactoring: enable chain_store to perform all the transaction validation where chain is currently used
This is to simplify parallelization of ingress handling, removing the need to store the `chain` and fall-back to using the chain_store.
