+++ chain/chain/src/store/mod.rs
@@ -47,7 +47,6 @@ use near_store::{
     KeyForStateChanges, LARGEST_TARGET_HEIGHT_KEY, LATEST_KNOWN_KEY, PartialStorage, Store,
     StoreUpdate, TAIL_KEY, WrappedTrieChanges,
 };
-use utils::get_block_header_on_chain_by_height;
 
 use crate::types::{Block, BlockHeader, LatestKnown};
 use near_store::db::{STATE_SYNC_DUMP_KEY, StoreStatistics};
@@ -492,57 +491,11 @@ impl ChainStore {
         prev_block_header: &BlockHeader,
         base_block_hash: &CryptoHash,
     ) -> Result<(), InvalidTxError> {
-        // if both are on the canonical chain, comparing height is sufficient
-        // we special case this because it is expected that this scenario will happen in most cases.
-        let base_height =
-            self.get_block_header(base_block_hash).map_err(|_| InvalidTxError::Expired)?.height();
-        let prev_height = prev_block_header.height();
-        if let Ok(base_block_hash_by_height) = self.get_block_hash_by_height(base_height) {
-            if &base_block_hash_by_height == base_block_hash {
-                if let Ok(prev_hash) = self.get_block_hash_by_height(prev_height) {
-                    if &prev_hash == prev_block_header.hash() {
-                        if prev_height <= base_height + self.transaction_validity_period {
-                            return Ok(());
-                        } else {
-                            return Err(InvalidTxError::Expired);
-                        }
-                    }
-                }
-            }
-        }
-
-        // if the base block height is smaller than `last_final_height` we only need to check
-        // whether the base block is the same as the one with that height on the canonical fork.
-        // Otherwise we walk back the chain to check whether base block is on the same chain.
-        let last_final_height = self
-            .get_block_height(prev_block_header.last_final_block())
-            .map_err(|_| InvalidTxError::InvalidChain)?;
-
-        if prev_height > base_height + self.transaction_validity_period {
-            Err(InvalidTxError::Expired)
-        } else if last_final_height >= base_height {
-            let base_block_hash_by_height = self
-                .get_block_hash_by_height(base_height)
-                .map_err(|_| InvalidTxError::InvalidChain)?;
-            if &base_block_hash_by_height == base_block_hash {
-                if prev_height <= base_height + self.transaction_validity_period {
-                    Ok(())
-                } else {
-                    Err(InvalidTxError::Expired)
-                }
-            } else {
-                Err(InvalidTxError::InvalidChain)
-            }
-        } else {
-            let header =
-                get_block_header_on_chain_by_height(self, prev_block_header.hash(), base_height)
-                    .map_err(|_| InvalidTxError::InvalidChain)?;
-            if header.hash() == base_block_hash {
-                Ok(())
-            } else {
-                Err(InvalidTxError::InvalidChain)
-            }
-        }
+        self.store.check_transaction_validity_period(
+            prev_block_header,
+            base_block_hash,
+            self.transaction_validity_period,
+        )
     }
 
     pub fn compute_transaction_validity(
